Fixed some problems
