+++ simplq/src/components/common/Button/index.jsx
@@ -1,5 +1,9 @@
-import StandardButton from './Button';
-import { CreateQueueButton } from './Button.stories';
+import StandardButton, {
+  CreateQueueButton,
+  ShareButton,
+  RefreshButton,
+  ForkOnGithubButton,
+} from './Button';
 
-export { CreateQueueButton };
+export { CreateQueueButton, ShareButton, RefreshButton, ForkOnGithubButton };
 export default StandardButton;
