+++ simplq/src/components/common/Footer/index.jsx
@@ -5,7 +5,7 @@ import YouTubeIcon from '@material-ui/icons/YouTube';
 import GitHubIcon from '@material-ui/icons/GitHub';
 import MailIcon from '@material-ui/icons/Mail';
 import styles from './footer.module.scss';
-import { ForkOnGithubButton } from '../Button/Button.stories';
+import { ForkOnGithubButton } from '../Button';
 import ClickableLogo from '../ClickableLogo';
 
 export default () => (
