+++ simplq/src/components/common/Nav/Nav.stories.jsx
@@ -1,8 +1,10 @@
-import Nav from './LeftNav';
+import React from 'react';
+import Nav from './Burger';
 
 export default {
   component: Nav,
   title: 'Nav',
 };
 
-export const MainNav = Nav;
+export const HomeNav = () => <Nav page="home" />;
+export const AdminNav = () => <Nav page="admin" />;
